% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{greta.censored}
\alias{greta.censored-package}
\alias{greta.censored}
\title{Censored Distributions for 'greta'}
\description{
Provides additional censored distributions for use with Greta,
a probabilistic programming framework for Bayesian modeling. Includes
censored versions of Normal, Log-Normal, Student's t, Gamma, Exponential,
Weibull, Pareto, and Beta distributions with support for right, left,
and interval censoring.
}
\examples{

# add a simple example here to introduce the package!

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mtwesley/greta.censored}
  \item Report bugs at \url{https://github.com/mtwesley/greta.censored/issues}
}

}
\author{
\strong{Maintainer}: Mlen-Too Wesley \email{mlen.too.wesley@gmail.com}

}
