% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{greta.template}
\alias{greta.template-package}
\alias{greta.template}
\title{What the Package Does (One Line, Title Case)}
\description{
describe your package here, you can re-use the text from DESCRIPTION
}
\examples{

# add a simple example here to introduce the package!

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mtwesley/greta.censored}
  \item Report bugs at \url{https://github.com/mtwesley/greta.censored/issues}
}

}
\author{
\strong{Maintainer}: Mlen-Too Wesley \email{mlen.too.wesley@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
